+++ deps/libiosuhax/.gitignore
@@ -1,0 +1,5 @@
+/*.a
+/build
+*.bz2
+release/
+lib/
