+++ deps/libiosuhax/Dockerfile
@@ -1,0 +1,9 @@
+FROM wiiuenv/devkitppc:20210101
+
+WORKDIR tmp_build
+COPY . .
+RUN make clean && make && mkdir -p /artifacts/wut/usr && cp -r lib /artifacts/wut/usr && cp -r include /artifacts/wut/usr
+WORKDIR /artifacts
+
+FROM scratch
+COPY --from=0 /artifacts /artifacts
