+++ deps/libiosuhax/README.md
@@ -1,0 +1,10 @@
+[![Build Status](https://travis-ci.org/dimok789/libiosuhax.svg)](https://travis-ci.org/dimok789/libiosuhax)
+# libiosuhax
+A PPC library to access IOSUHAX from PPC and a devoptab for any device or path. 
+It's only compatible to RPX-Files.
+
+## Building
+Make you to have [wut](https://github.com/devkitPro/wut/) installed and use the following command for build:
+```
+make install
+```
